[Udhay] Extentions Implemented on all variables

+++ C0752248_MID_TERM_MAD3004/C0752248_MID_TERM_MAD3004/Extention/Gb.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-extension Int{
+extension Float{
     public func gigaByte() -> String{
         return String.init(format: "%d GB")
     }
