+++ C0752248_MID_TERM_MAD3004/C0752248_MID_TERM_MAD3004/Extention/Units.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-extension Int{
+extension Float{
     public func units() -> String{
         return String.init(format: "%d Units")
     }
